Rename parameter.rb to parameter_model.rb and remove separate extension
